+++ src/SearchOptions.cs
@@ -33,4 +33,9 @@ public class SearchOptions
     /// Kind option the source (optional)
     /// </summary>
     public string? SourceKind { get; set; }
+
+    /// <summary>
+    /// The Threshold Score to return (Not called min/max as some vector stores see low numbers as most similar while others see high numbers as most similar. The system will check that on the fly and use the indicated score as min/max)
+    /// </summary>
+    public double? ThresholdSimilarityScoreToReturn { get; set; }
 }